# Direct implementation of is_metavar_name and `is_metavar_ellipsis`, not via memoized regex